web: wip fixing topbar
I'm seeing two more problems that i kinda want to fix at the same
time:

1. Typing '?' while the search bar is focused toggles the keyboard
help on the dashboard
2. Selecting a search query via the dropdown items (like 'team:'
or 'status:' does not update the URL)

I'm starting to have a bad feeling (the term 'shared mutable state'
comes to mind) about having multiple `handleCallback` or
`handleDelivery` functions affecting the same data...

This might *also* be the right time to pull the TopBar logic apart.
Probably only the non-HD dashboard (in the non-zero-pipelines case)
should display a search bar, and the only 'page-independent' data
is the stuff relating to the user menu.

accordingly, maybe the breadcrumb logic should be pulled down
and duplicated across the various pages that need breadcrumbs,
and the pin icon should the the pipeline page's responsibility!

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>